pyproject.toml
[project]
name = "ariel"
version = "0.0.0"
description = "ARIEL"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "GPL-3.0" }
authors = [{ name = "jmdm", email = "jmdm.880@gmail.com" }]
classifiers = ["Development Status :: 2 - Pre-Alpha"]

dependencies = [
  "attrs>=25.4.0",
  "deap>=1.4.3",
  "evotorch>=0.6.1",
  "gymnasium>=1.2.3",
  "ipykernel>=6.30.1",
  "jupyter>=1.1.1",
  "kaleido>=1.2.0",
  "keyboard>=0.13.5",
  "manifold3d>=3.4.0",
  "matplotlib>=3.9.4",
  "mujoco>=3.3.6",
  "networkx>=3.2.1",
  "nevergrad>=1.0.12",
  "nicegui>=2.24.1",
  "numpy>=2.0.2",
  "numpy-quaternion>=2023.0.3",
  "opencv-python>=4.11.0.86",
  "panel>=1.8.9",
  "pillow>=11.3.0",
  "plotly>=6.6.0",
  "polars>=1.35.2",
  "pydantic>=2.11.9",
  "pydantic-settings>=2.10.1",
  "pyserial>=3.5",
  "pyvista>=0.47.1",
  "rich>=14.1.0",
  "sqlalchemy>=2.0.43",
  "sqlmodel>=0.0.25",
  "torch>=2.8.0",
  "torchvision>=0.23.0",
]


[project.urls]
Homepage = "https://github.com/ci-group/ariel"
Repository = "https://github.com/ci-group/ariel"
Documentation = "https://ariel.readthedocs.io"
Changelog = "https://github.com/ci-group/ariel/releases"

[dependency-groups]
dev = ["mypy>=1.18.2", "pytest>=8.4.2", "pytest-modern>=0.7.4", "ruff>=0.13.3"]
docs = [
  "jupyter-sphinx>=0.5.3",
  "linkify-it-py>=2.0.3",
  "myst-parser>=3.0.1",
  "nbsphinx>=0.9.7",
  "nox>=2025.5.1",
  "numpydoc>=1.9.0",
  "pandoc>=2.4",
  "pypandoc>=1.15",
  "shibuya>=2025.8.16",
  "sphinx-autoapi>=3.6.0",
  "sphinx-autobuild>=2024.10.3",
  "sphinx-autodoc-typehints>=3.0.1",
  "sphinx-autodoc2>=0.5.0",
  "sphinx-click>=6.0.0",
  "sphinx-copybutton>=0.5.2",
  "sphinx-design>=0.6.1",
]
stubs = [
  "lxml-stubs>=0.5.1",
  "opencv-stubs>=0.1.0",
  "pandas-stubs>=2.3.2.250926",
  "scipy-stubs>=1.16.2.0",
  "types-aiofiles>=24.1.0.20250822",
  "types-colorama>=0.4.15.20250801",
  "types-decorator>=5.2.0.20250324",
  "types-defusedxml>=0.7.0.20250822",
  "types-docutils>=0.22.2.20250924",
  "types-greenlet>=3.2.0.20250915",
  "types-jsonschema>=4.25.1.20250822",
  "types-libsass>=0.23.0.20250401",
  "types-networkx>=3.5.0.20250918",
  "types-openpyxl>=3.1.5.20250919",
  "types-pexpect>=4.9.0.20250916",
  "types-protobuf>=6.32.1.20250918",
  "types-pygments>=2.19.0.20250809",
  "types-pyyaml>=6.0.12.20250915",
  "types-setuptools>=80.9.0.20250822",
  "types-tabulate>=0.9.0.20241207",
  "types-tqdm>=4.67.0.20250809",
  "types-ujson>=5.10.0.20250822",
  "types-xlrd>=2.0.0.20250914",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.uv]
package = true

[project.scripts]
ariel = "ariel.__main__:cli"

[build-system]
# When using mypyc, you must list all dependencies used by the compiled
# modules here. Otherwise, mypyc will not be able to see the types of the
# dependencies, and the compilation will fail.
requires = ["setuptools>=61.0", "mypy>=1.0"]
build-backend = "setuptools.build_meta"

[tool.coverage.paths]
source = ["src", "*/site-packages"]
tests = ["tests", "*/tests"]

[tool.coverage.run]
branch = true
source = ["ariel", "tests"]

[tool.coverage.report]
show_missing = true
fail_under = 100
exclude_lines = ["pragma: no cover", "if TYPE_CHECKING:"]

[tool.pydoclint]
style = 'numpy'
exclude = '\.git|\.nox|noxfile.py'
arg-type-hints-in-docstring = false
